@@ -5,7 +5,6 @@ module Dovado
     class Sms
       class Message
         include Celluloid
-        include Celluloid::Logger
         
         attr_reader :id, :body, :pdus, :from, :sent, :encoding
   
@@ -25,7 +24,6 @@ module Dovado
             @sent = args[:sent] unless args[:sent].nil?
             @encoding = args[:encoding] unless args[:encoding].nil?
           end
-          debug "Starting up #{self.class.to_s}..."
         end
   
         def self.from_string string=nil
@@ -37,25 +35,28 @@ module Dovado
             if row_array.length == 2
               key = row_array[0].downcase
               val = row_array[1]
-              case key
+              case key.strip
               when 'from'
-                hash[:from] = val
-              when 'sent'
-                hash[:sent] = DateTime.parse(val)
+                hash[:from] = val.strip
               when 'alphabet'
-                hash[:encoding] = val
+                begin
+                  hash[:encoding] = Encoding.find(val.strip)
+                rescue ArgumentError
+                  hash[:encoding] = Encoding::UTF_8
+                end
               when 'id'
-                hash[:id] = val
-              else
-                # do nothing
+                hash[:id] = val.strip.to_i
               end
-        
-            end
-            unless row.downcase =~ /end of sms/
-              message_body += "#{row}\n"
+            elsif row_array.length > 2
+              sent = row.match(/[Ss]ent\:\W(.*)/)
+              hash[:sent] = DateTime.parse(sent[1].strip)
+            else
+              unless row.downcase =~ /end of sms/
+                message_body += "#{row}\n"
+              end
             end
           end
-          hash[:body] = message_body
+          hash[:body] = message_body.tr(">>", "").tr("\x17", "").strip.force_encoding(hash[:encoding])
     
           Dovado::Router::Sms::Message.new(hash)
         end
